fix: Move the notifications utils outside the pages.
